feat(certus-grpc-connector): target vLLM 0.20.0; add Docker test harness
Verified the connector against vLLM 0.20.0's v1 kv_offload API and fixed
compatibility gaps found in review:
- GPULoadStoreSpec.block_ids is an np.int64 array in 0.20.0; cast block ids to
  int before offset math and protobuf, so IpcHandle.offset (uint64) accepts
  them. Updated the test's fake GPULoadStoreSpec to match (numpy array,
  3-arg constructor).
- Pinned the targeted vLLM version in pyproject.toml (0.20.0, provided by the
  runtime env).

Add a container-based test path for the 450-conv / 12-turn workload:
- run_multiturn_grpc_certus.py: ShareGPT multi-turn harness driving
  CertusGrpcOffloadingSpec against a running certus-server (CERTUS_SERVER env),
  defaulting to the 450-conv/12-turn dataset.
- Dockerfile on docker.io/vllm/vllm-openai:v0.20.0 (fully-qualified for
  docker+podman) installing the connector; runs the workload as CMD. The
  certus-server runs on the host (SPDK/NVMe/hugepages), reached over gRPC with
  --ipc=host for CUDA IPC.

+++ certus-grpc-connector/certus_grpc_connector/gpu.py
@@ -60,7 +60,9 @@ class KvCacheIpc:
     base_delta: int  # data_ptr - alloc_base
 
     def block_offset(self, block_id: int) -> int:
-        return self.base_delta + block_id * self.stride_bytes
+        # int() guards against numpy int64 block ids (GPULoadStoreSpec stores
+        # block_ids as an np.int64 array), which protobuf's uint64 field rejects.
+        return int(self.base_delta + int(block_id) * self.stride_bytes)
 
 
 def current_device() -> int:
